Cast `rc` to uint32_t to mute a -Wsign-compare warning

+++ usr.sbin/bsnmpd/tools/bsnmptools/bsnmpget.c
@@ -500,7 +500,7 @@ snmptool_walk(struct snmp_toolinfo *snmptoolctx)
 			outputs += rc;
 			snmp_pdu_free(&resp);
 
-			if (rc < resp.nbindings)
+			if ((uint32_t)rc < resp.nbindings)
 				break;
 
 			snmpwalk_nextpdu_create(op,
